+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Batch` prints total estimated cost if `verbose=True`.
 - Unified config and authentication.
 - Remove restriction that `JaxCustomMedium` must not be a 3D pixelated array.
+- Limit total number of input structures in adjoint plugin for performance reasons.
 
 ### Fixed
 - Plotting 2D materials in `SimulationData.plot_field` and other circumstances.